+++ src/models/mod.rs
@@ -1,12 +1,12 @@
-pub mod sprite;
-pub mod layer_tile;
 pub mod layer_header;
+pub mod layer_tile;
+pub mod sprite;
 
+use crate::constants::TILES_PER_LAYER;
 use std::io;
 use std::io::ErrorKind;
-use crate::constants::{TILES_PER_LAYER_COLUMN, TILES_PER_LAYER_ROW};
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Default, Eq, PartialEq)]
 pub struct Sprite {
     pub x: usize,
     pub y: usize,
@@ -25,10 +25,16 @@ pub struct Sprite {
 #[derive(Debug, Clone)]
 pub struct Layer {
     pub header: LayerHeader,
-    pub content: [LayerTile; TILES_PER_LAYER_ROW  * TILES_PER_LAYER_COLUMN]
+    pub content: [LayerTile; TILES_PER_LAYER],
 }
 
-#[derive(Debug, Clone)]
+impl Layer {
+    pub fn new(header: LayerHeader, content: [LayerTile; TILES_PER_LAYER]) -> Self {
+        Self { header, content }
+    }
+}
+
+#[derive(Debug, Clone, Default, Eq, PartialEq)]
 pub struct LayerHeader {
     pub x: isize,
     pub y: isize,
@@ -37,22 +43,25 @@ pub struct LayerHeader {
     pub order: usize,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Default, Eq, PartialEq)]
 pub struct LayerTile {
     pub id: usize,
     pub flip_v: bool,
     pub flip_h: bool,
     pub palette: usize,
     pub half_alpha: bool,
-    pub rotated: bool
+    pub rotated: bool,
 }
 
 pub trait Byteable {
     const SIZE: usize;
 
-    fn try_from_bytes(bytes: &[u8]) -> Result<Self, io::Error> where Self: Sized {
+    fn try_from_bytes(bytes: &[u8]) -> Result<Self, io::Error>
+    where
+        Self: Sized,
+    {
         if bytes.len() < Self::SIZE {
-            return Err(io::Error::from(ErrorKind::UnexpectedEof))
+            Err(io::Error::from(ErrorKind::UnexpectedEof))
         } else {
             Ok(Self::from_bytes(bytes))
         }